Refactor settings into multiple real pages
